requirements.txt
numpy>=1.18.5
pillow>=9.0.0
tensorflow>=2.10.0
streamlit>=1.0.0
openai>=0.27.0
